add modified code

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: aoshinth <aoshinth@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 21:39:39 by aoshinth          #+#    #+#             */
-/*   Updated: 2024/04/17 16:04:04 by aoshinth         ###   ########.fr       */
+/*   Updated: 2024/04/25 15:40:52 by aoshinth         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
 	int	i;
